pyproject.toml
[tool.poetry]
name = "sussllm"
version = "0.1.0"
description = ""
authors = ["Saber Zerhoudi <saber.zerhoudi@uni-passau.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.4"
pandas = "^2.2.2"
scikit-learn = "^1.5.0"
tensorflow = "^2.16.1"
transformers = "^4.41.2"
pytest = "^8.2.1"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
openai = "^1.30.5"
tqdm = "^4.66.4"
flask = "^3.0.3"
beautifulsoup4 = "^4.12.3"
progressbar = "^2.5"
rank-bm25 = "^0.2.2"
bert-score = "^0.3.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"